+++ vm.c
@@ -9,7 +9,7 @@ void vm_collect_garbage(vm_t *vm) {
 }
 
 void sweep(vm_t *vm) {
-  for (int i = 0; i < vm->objects->count; i++) {
+  for (size_t i = 0; i < vm->objects->count; i++) {
     snek_object_t *obj = vm->objects->data[i];
     if (obj->is_marked) {
       obj->is_marked = false;
@@ -39,7 +39,7 @@ void trace(vm_t *vm) {
   }
 
   // Get previously marked objects (which are the roots)
-  for (int i = 0; i < vm->objects->count; i++) {
+  for (size_t i = 0; i < vm->objects->count; i++) {
     snek_object_t *obj = vm->objects->data[i];
     if (obj->is_marked) {
       stack_push(gray_objects, obj);
@@ -105,13 +105,13 @@ vm_t *vm_new() {
 
 void vm_free(vm_t *vm) {
   // Free the stack frames, and then their container
-  for (int i = 0; i < vm->frames->count; i++) {
+  for (size_t i = 0; i < vm->frames->count; i++) {
     frame_free(vm->frames->data[i]);
   }
   stack_free(vm->frames);
 
   // Free the objects, and then their container
-  for (int i = 0; i < vm->objects->count; i++) {
+  for (size_t i = 0; i < vm->objects->count; i++) {
     snek_object_free(vm->objects->data[i]);
   }
   stack_free(vm->objects);
